Handle ecdsa with refactor
