+++ AmazonSecretsManagerHandler/AmazonSecretsManagerHandler.csproj
@@ -9,7 +9,6 @@
   <ItemGroup>
     <PackageReference Include="AWSSDK.SecretsManager" Version="3.7.400.130" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-    <ProjectReference Include="../SignatureHandling/SignatureHandling.csproj"/> 
   </ItemGroup>
 
 </Project>
