+++ CoinbaseWebSocketConstructs/CoinbaseWebSocketConstructs.csproj
@@ -12,6 +12,7 @@
    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="8.8.0" />
    <ProjectReference Include="../AmazonSecretsManagerHandler/AmazonSecretsManagerHandler.csproj" />
+   <ProjectReference Include="../SignatureHandling/SignatureHandling.csproj" />
   </ItemGroup>
 
 </Project>
