+++ SignatureHandling/SignatureAlgorithmFactory.cs
@@ -1,16 +1,17 @@
 namespace SignatureHandling;
 
 using SignatureHandling.Interfaces;
+using AmazonSecretsManagerHandler;
 
 public static class SignatureAlgorithmFactory
 {
-    public static ISignatureAlgorithm Create(SigningMetadata metadata)
+    public static ISignatureAlgorithm Create()
     {
-        return metadata.Algorithm switch
+        return SecretsProvider.GetAlgorithmString() switch
         {
-            "ecdsa" => new EcdsaSignatureAlgorithm(metadata.Secret, metadata.KeyId),
-            "ed25519" => new Ed25519SignatureAlgorithm(metadata.Secret, metadata.KeyId),
-            _ => throw new NotSupportedException($"Unsupported algorithm: {metadata.Algorithm}")
+            "ecdsa" => new EcdsaSignatureAlgorithm(SecretsProvider.GetSecretKey(), SecretsProvider.GetApiKeyName()),
+            "ed25519" => new Ed25519SignatureAlgorithm(SecretsProvider.GetSecretKey(), SecretsProvider.GetApiKeyName()),
+            _ => throw new NotSupportedException($"Unsupported algorithm: {SecretsProvider.GetAlgorithmString()}")
         };
     }
 }
